missing cast in AccessLOg

+++ hadoop-hdfs-project/hadoop-hdfs/src/main/java/org/apache/hadoop/hdfs/server/namenode/AccessLog.java
@@ -89,7 +89,7 @@ public class AccessLog extends Configured implements Runnable {
       } catch (URISyntaxException | IOException | ParseException e) {
         LOG.error("Exception while running access log scan", e);
       }
-      Thread.sleep(interval * 3600 * 1000);
+      Thread.sleep((long) (interval * 3600 * 1000));
     }
   }
   public void runScan() throws URISyntaxException, IOException, MalformedURLException, ParseException {
